+++ actionpack/test/template/number_helper_i18n_test.rb
@@ -47,18 +47,19 @@ class NumberHelperTest < ActionView::TestCase
     assert_equal("-10.00 - &$", number_to_currency(-10, :locale => 'ts', :format => "%n - %u"))
   end
 
-  def test_number_to_currency_with_clean_i18n_settings
-    clean_i18n do
-      assert_equal("$10.00", number_to_currency(10))
-      assert_equal("-$10.00", number_to_currency(-10))
-    end
+  def test_number_to_currency_with_empty_i18n_store
+    I18n.backend.store_translations 'empty', {}
+
+    assert_equal("$10.00", number_to_currency(10, :locale => 'empty'))
+    assert_equal("-$10.00", number_to_currency(-10, :locale => 'empty'))
   end
-  
+
   def test_number_to_currency_without_currency_negative_format
-    clean_i18n do
-      I18n.backend.store_translations 'ts', :number => { :currency => { :format => { :unit => '@', :format => '%n %u' } } }
-      assert_equal("-10.00 @", number_to_currency(-10, :locale => 'ts'))
-    end
+    I18n.backend.store_translations 'no_negative_format', :number => {
+      :currency => { :format => { :unit => '@', :format => '%n %u' } }
+    }
+
+    assert_equal("-10.00 @", number_to_currency(-10, :locale => 'no_negative_format'))
   end
 
   def test_number_with_i18n_precision
@@ -67,21 +68,24 @@ class NumberHelperTest < ActionView::TestCase
 
     #Precision inherited and significant was set
     assert_equal("1.00", number_with_precision(1.0, :locale => 'ts'))
-
   end
 
-  def test_number_with_empty_i18n_store
+  def test_number_with_i18n_precision_and_empty_i18n_store
     I18n.backend.store_translations 'empty', {}
 
-    #Precision should be defaulted to 3, separator to ".", delimiter to ""
     assert_equal("123456789.123", number_with_precision(123456789.123456789, :locale => 'empty'))
-    #Significant and strip_insignificant_zeros should be defaulted to false
     assert_equal("1.000", number_with_precision(1.0000, :locale => 'empty'))
   end
 
   def test_number_with_i18n_delimiter
-    #Delimiter "," and separator "."
+    # Delimiter "," and separator "."
     assert_equal("1,000,000.234", number_with_delimiter(1000000.234, :locale => 'ts'))
+  end
+
+  def test_number_with_i18n_delimiter_and_empty_i18n_store
+    I18n.backend.store_translations 'empty', {}
+
+    assert_equal("1,000,000.234", number_with_delimiter(1000000.234, :locale => 'empty'))
   end
 
   def test_number_to_i18n_percentage
@@ -93,10 +97,25 @@ class NumberHelperTest < ActionView::TestCase
     assert_equal("12434%", number_to_percentage(12434, :locale => 'ts'))
   end
 
+  def test_number_to_i18n_percentage_and_empty_i18n_store
+    I18n.backend.store_translations 'empty', {}
+
+    assert_equal("1.000%", number_to_percentage(1, :locale => 'empty'))
+    assert_equal("1.243%", number_to_percentage(1.2434, :locale => 'empty'))
+    assert_equal("12434.000%", number_to_percentage(12434, :locale => 'empty'))
+  end
+
   def test_number_to_i18n_human_size
     #b for bytes and k for kbytes
     assert_equal("2 k", number_to_human_size(2048, :locale => 'ts'))
     assert_equal("42 b", number_to_human_size(42, :locale => 'ts'))
+  end
+
+  def test_number_to_i18n_human_size_with_empty_i18n_store
+    I18n.backend.store_translations 'empty', {}
+
+    assert_equal("2 KB", number_to_human_size(2048, :locale => 'empty'))
+    assert_equal("42 Bytes", number_to_human_size(42, :locale => 'empty'))
   end
 
   def test_number_to_human_with_default_translation_scope
@@ -113,19 +132,15 @@ class NumberHelperTest < ActionView::TestCase
     assert_equal "2 Tens", number_to_human(20, :locale => 'ts')
   end
 
+  def test_number_to_human_with_empty_i18n_store
+    I18n.backend.store_translations 'empty', {}
+
+    assert_equal "2 Thousand", number_to_human(2000, :locale => 'empty')
+    assert_equal "1.23 Billion", number_to_human(1234567890, :locale => 'empty')
+  end
+
   def test_number_to_human_with_custom_translation_scope
     #Significant was set to true with precision 2, with custom translated units
     assert_equal "4.3 cm", number_to_human(0.0432, :locale => 'ts', :units => :custom_units_for_number_to_human)
   end
-
-  private
-    def clean_i18n
-      load_path = I18n.load_path.dup
-      I18n.load_path.clear
-      I18n.reload!
-      yield
-    ensure
-      I18n.load_path = load_path
-      I18n.reload!
-    end
 end